Move code samples to MainActivity
